[1827] DRY up editing basic detail controllers
Move shared methods into EditBasicDetail concern
